atualizado comments controller com autenticação

+++ src/controllers/CommentsController.js
@@ -1,13 +1,24 @@
 const knex = require("../database/knex");
+const AppError = require("../utils/AppError");
 
 class CommentsController {
     async create(request, response) {
         const { content } = request.body;
         const { post_id } = request.params;
+        const user_id = request.user.id
+        console.log(post_id)
+
+
+        const [originalPost] = await knex("posts").where({ id: post_id });
+
+        if (!originalPost) {
+            throw new AppError("Post não encontrado")
+        }
 
         await knex("comments").insert({
             content,
-            post_id
+            post_id,
+            user_id
         });
 
         response.json();
@@ -23,6 +34,13 @@ class CommentsController {
 
     async delete(request, response) {
         const { id } = request.params;
+        const user_id = request.user.id;
+
+        const [comment] = await knex("comments").where({ id });
+
+        if(comment.user_id !== user_id) {
+            throw new AppError("Você só pode deletar comentários seus!")
+        }
 
         await knex("comments").where({ id }).delete();
